refactor: move ciphersuites to a unified module
