+++ crypto-ffi/src/wasm/mod.rs
@@ -25,96 +25,7 @@ use utils::*;
 use wasm_bindgen::{JsCast, prelude::*};
 use wasm_bindgen_futures::future_to_promise;
 
-use crate::{CoreCryptoError, InternalError, MlsError, WasmCryptoResult};
-
-#[allow(non_camel_case_types)]
-#[wasm_bindgen]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize, strum::FromRepr)]
-#[repr(u16)]
-// see [core_crypto::prelude::CiphersuiteName]
-pub enum Ciphersuite {
-    /// DH KEM x25519 | AES-GCM 128 | SHA2-256 | Ed25519
-    MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519 = 0x0001,
-    /// DH KEM P256 | AES-GCM 128 | SHA2-256 | EcDSA P256
-    MLS_128_DHKEMP256_AES128GCM_SHA256_P256 = 0x0002,
-    /// DH KEM x25519 | Chacha20Poly1305 | SHA2-256 | Ed25519
-    MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 = 0x0003,
-    /// DH KEM x448 | AES-GCM 256 | SHA2-512 | Ed448
-    MLS_256_DHKEMX448_AES256GCM_SHA512_Ed448 = 0x0004,
-    /// DH KEM P521 | AES-GCM 256 | SHA2-512 | EcDSA P521
-    MLS_256_DHKEMP521_AES256GCM_SHA512_P521 = 0x0005,
-    /// DH KEM x448 | Chacha20Poly1305 | SHA2-512 | Ed448
-    MLS_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 = 0x0006,
-    /// DH KEM P384 | AES-GCM 256 | SHA2-384 | EcDSA P384
-    MLS_256_DHKEMP384_AES256GCM_SHA384_P384 = 0x0007,
-}
-
-impl From<MlsCiphersuite> for Ciphersuite {
-    fn from(c: MlsCiphersuite) -> Self {
-        Self::from(CiphersuiteName::from(c))
-    }
-}
-
-impl From<Ciphersuite> for MlsCiphersuite {
-    fn from(c: Ciphersuite) -> Self {
-        let c: CiphersuiteName = c.into();
-        Self::from(c)
-    }
-}
-
-impl From<CiphersuiteName> for Ciphersuite {
-    fn from(c: CiphersuiteName) -> Self {
-        match c {
-            CiphersuiteName::MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519 => {
-                Self::MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-            }
-            CiphersuiteName::MLS_128_DHKEMP256_AES128GCM_SHA256_P256 => Self::MLS_128_DHKEMP256_AES128GCM_SHA256_P256,
-            CiphersuiteName::MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-                Self::MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519
-            }
-            CiphersuiteName::MLS_256_DHKEMX448_AES256GCM_SHA512_Ed448 => Self::MLS_256_DHKEMX448_AES256GCM_SHA512_Ed448,
-            CiphersuiteName::MLS_256_DHKEMP521_AES256GCM_SHA512_P521 => Self::MLS_256_DHKEMP521_AES256GCM_SHA512_P521,
-            CiphersuiteName::MLS_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-                Self::MLS_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448
-            }
-            CiphersuiteName::MLS_256_DHKEMP384_AES256GCM_SHA384_P384 => Self::MLS_256_DHKEMP384_AES256GCM_SHA384_P384,
-        }
-    }
-}
-
-#[allow(clippy::from_over_into)]
-impl Into<CiphersuiteName> for Ciphersuite {
-    fn into(self) -> CiphersuiteName {
-        match self {
-            Self::MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519 => {
-                CiphersuiteName::MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-            }
-            Self::MLS_128_DHKEMP256_AES128GCM_SHA256_P256 => CiphersuiteName::MLS_128_DHKEMP256_AES128GCM_SHA256_P256,
-            Self::MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519 => {
-                CiphersuiteName::MLS_128_DHKEMX25519_CHACHA20POLY1305_SHA256_Ed25519
-            }
-            Self::MLS_256_DHKEMX448_AES256GCM_SHA512_Ed448 => CiphersuiteName::MLS_256_DHKEMX448_AES256GCM_SHA512_Ed448,
-            Self::MLS_256_DHKEMP521_AES256GCM_SHA512_P521 => CiphersuiteName::MLS_256_DHKEMP521_AES256GCM_SHA512_P521,
-            Self::MLS_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448 => {
-                CiphersuiteName::MLS_256_DHKEMX448_CHACHA20POLY1305_SHA512_Ed448
-            }
-            Self::MLS_256_DHKEMP384_AES256GCM_SHA384_P384 => CiphersuiteName::MLS_256_DHKEMP384_AES256GCM_SHA384_P384,
-        }
-    }
-}
-
-/// Helper to lower arrays of Ciphersuites (js -> rust)
-fn lower_ciphersuites(ciphersuites: &[u16]) -> WasmCryptoResult<Vec<MlsCiphersuite>> {
-    ciphersuites.iter().try_fold(
-        Vec::with_capacity(ciphersuites.len()),
-        |mut acc, &cs| -> WasmCryptoResult<_> {
-            let cs = Ciphersuite::from_repr(cs).ok_or(InternalError::UnknownCiphersuite)?;
-            let cs: MlsCiphersuite = cs.into();
-            acc.push(cs);
-            Ok(acc)
-        },
-    )
-}
+use crate::{Ciphersuite, CoreCryptoError, InternalError, MlsError, WasmCryptoResult, lower_ciphersuites};
 
 #[allow(non_camel_case_types)]
 #[wasm_bindgen]
